backend/pyproject.toml
[project]
name = "smartqueue-backend"
version = "0.1.0"
description = "Backend Django pour SmartQueue (multi-tenant, API REST, WebSockets, tâches async)."
readme = "README.md"
authors = [
  { name = "SmartQueue Team" }
]
requires-python = ">=3.11"
dependencies = [
  "Django>=4.2,<5.0",
  "djangorestframework>=3.14",
  "djangorestframework-simplejwt>=5.3",
  "django-filter>=24.2",
  "django-cors-headers>=4.3",
  "channels>=4.0",
  "channels-redis>=4.1",
  "celery>=5.3",
  "redis>=5.0",
  "psycopg2-binary>=2.9",
  "python-dotenv>=1.0",
  "drf-spectacular>=0.27",
  "django-environ>=0.11",
  "structlog>=24.1",
  "twilio>=9.0",
  "sendgrid>=6.11",
  "firebase-admin>=6.4"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-django>=4.8",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5.0",
  "pytest-mock>=3.14",
  "model-bakery>=1.19",
  "black>=24.4",
  "ruff>=0.4",
  "mypy>=1.9",
  "django-extensions>=3.2",
  "freezegun>=1.5"
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["apps*", "smartqueue_backend*"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "smartqueue_backend.settings.test"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--reuse-db",
    "--nomigrations",
    "--cov=apps",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=30",
]
testpaths = ["apps"]

[tool.coverage.run]
source = ["apps"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/__init__.py",
    "*/admin.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
